Ticket 29: Optimization / Performance on long lists Changed XJDocument When the user panel is opened the combo box calls numberOfItemsInComboBox.
Each character typed in then causes the combo box to call objectValueForItemAtIndex as many times as was returned from numberOfItemsInComboBox.

Each call to objectValueForItemAtIndex was doing a lookup in LJKit to the current logged in acct's user list. For users with 250 friends, 2 characters typed totalled to 500 trips to this LJKit call.

Now when numberOfItemsInComboBox is called the user or community array is looked up once and stored locally into XJDocument. This way the 500 calls do not require 500 LJKit lookup, just 500 compares of a local array. Diffrence is very noticable.


git-svn-id: svn://svn.code.sf.net/p/ljkit/code/Xjournal/trunk@141 553375af-a522-0410-a719-e18aa8ed40dd